% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_identities.R
\name{distance_identities}
\alias{distance_identities}
\title{Decay of identity along a transect}
\usage{
distance_identities(transects, positions, labels = 1:length(positions))
}
\arguments{
\item{transects}{List of transects. Each element in the list should be a 1-d
vector with an element for every sampling point along the transect, with each
element the genotype observed at that sampliong point.}

\item{positions}{1-d vector of positions along the transect. This should be
the same length of each vector in \code{transects}.}

\item{labels}{Optional vector of labels for each sampling point. Defaults to
integers.}
}
\value{
Returns a data.frame with a row for each pair of sampling points.
Columns give the labels for the points being compared, their distance, the
number of matches between them, and the number of non-NA observations for
each pair.
}
\description{
Calculates the probability of observing identical genotypes at different
distances along a transect.
}
\details{
\code{distance_identities} calculates how often pairs of identical genotypes are
observed for all possible distances between sampling points along a transect.
Identities are averaged over multiple years for a transect.
}
\examples{
positions <- sort(runif(16, -5,5))
transects <- list(
  rep(letters[1:4], each=4),
    rep(letters[1:4], 4)
    )
distance_identities(transects, positions, labels = LETTERS[1:16])
}
\author{
Tom Ellis
}
