% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_clustering.R
\name{habitat_clustering}
\alias{habitat_clustering}
\title{Do genotypes cluster by microhabitat?}
\usage{
habitat_clustering(genotype, habitat)
}
\arguments{
\item{genotype}{Vector of genotype labels.}

\item{habitat}{Vector of habitat labels of the same length as \code{genotype}.}
}
\value{
Proportion of pairs of individuals in the same habitat that are
identical.
}
\description{
Calculates how many pairs of plants of the same genotype are also in the
same microhabitat.
}
\author{
Tom Ellis
}
