% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_population.R
\name{initialise_population}
\alias{initialise_population}
\title{Initialise population}
\usage{
initialise_population(
  mean_dispersal_distance,
  n_starting_genotypes,
  population_size,
  box_limit
)
}
\arguments{
\item{mean_dispersal_distance}{Float >0. Mean seed dispersal distance. The
reciprocal of this is used as the rate parameter to draw from the exponential
distribution.}

\item{n_starting_genotypes}{Int >0. Number of initial genotypes to start with.
Defaults to 50}

\item{population_size}{Integer > 1.}

\item{box_limit}{Float >1 giving half the circumference of the torus.
\code{sim_population} based on the length of the transect and plant density.}
}
\description{
Set up an initial population where each genotype starts from a single
spot and disperses from there.
}
