% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_table.R
\name{parameter_table}
\alias{parameter_table}
\title{Data.frame of parameter values}
\usage{
parameter_table(
  mean_dispersal_distance,
  outcrossing_rate,
  n_generations,
  n_starting_genotypes,
  density,
  nsims,
  n_sample_points,
  sample_spacing,
  range_limit,
  how_far_back,
  dormancy,
  habitat_width = (n_sample_points * sample_spacing * range_limit),
  population_size = density * habitat_width^2
)
}
\arguments{
\item{mean_dispersal_distance}{Float >0. Mean seed dispersal distance. The
reciprocal of this is used as the rate parameter to draw from the exponential
distribution.}

\item{outcrossing_rate}{Float between 0 and 1. Probability that an individual
is outcrossed.}

\item{n_generations}{Int >12. Number of generations to run the simulations.}

\item{n_starting_genotypes}{Int >0. Number of initial genotypes to start with.
Defaults to 50}

\item{density}{Float >0. Average density of plants per square metre.}

\item{nsims}{Int >0. Number of replicate populations to simulate.}

\item{n_sample_points}{Number of points to sample along the transect.}

\item{sample_spacing}{Distance between sampling points.}

\item{range_limit}{Float >1 defining how much wider than the transect the
width of the habitat should be. This, along with plant density, determines
how many plants will be simulated. Defaults to 1.5. Ignored if
\code{pop_structure="hardcoded"}

#' @param pop_structure Character string indicating the initial population
structure to be simulated. Passing "uniform" simulated a panmictic population.
"clusters" simulates clusters of identical individuals that disperse from
distinct mothers via exponential dispersal set by \code{mean_dispersal_distance}.
This is likely to generate very disperate clumps. Passing "mvnorm" simulates
uniformly distributed coordinates for indiduals, as well as centroid
positions for genotypes. Individuals are assigned a genotype in proportion to
their distance to each genotype centroid based on multivariate-normal
probabilities. The variance covariance matrix for this is set as
\code{sqrt(habitat_size/n_starting_genotypes) / 3} such that the tail of each
genotype just about touches those of its neighbours, on average.
If \code{pop_structure='hardcoded'} and a vector of genotypes is passed to
\code{n_starting_genotypes}, for example
observed genotypes from along all real-world transects, this simulates
bands of identical genotypes by copying the vector over an evenly-
spaced grid (giving horizontal but not vertical structure). There is one
round of dispersal from this initial generation via exponential dispersal
(controlled by \code{mixing}) and to get the population to the correct
population density.}

\item{how_far_back}{Integer number of generations to sample with a transect}

\item{dormancy}{Float between 0 and 1. Probability that a seedling is drawn
from the seed bank. Seedlings are drawn from the prior generation with
probability 1-dormancy.}

\item{habitat_width}{Length of the edge of the habitat square}

\item{population_size}{Number of individual plants in the population,
derived from the habitat size and population density}
}
\value{
A data.frame giving parameter names and values
}
\description{
Data.frame of parameter values
}
\author{
Tom Ellis
}
