% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_fitness.R
\name{relative_fitness}
\alias{relative_fitness}
\title{Relative fitness}
\usage{
relative_fitness(pop)
}
\arguments{
\item{pop}{A list describing genotypes and positions in a population.
Output of \code{initialise_population},}
}
\value{
A vector of relative fitness values.
}
\description{
Calculate relative fitness in the population under directional selection.
}
\details{
\code{relative_fitness} takes lof fitness values from a population generated by
\code{initialise_population} and exponentiates them to calculate relative fitness.
Relative fitness is absolute fitness divided by mean absolute fitness.

If individuals have NA phenotype the fitness is returned as zero.
}
