% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_population.R
\name{sim_population}
\alias{sim_population}
\title{Simulate a population}
\usage{
sim_population(
  grid_size,
  mean_dispersal_distance,
  outcrossing_rate,
  n_generations,
  n_starting_genotypes = 50,
  dormancy_rate = 0,
  verbose = TRUE,
  return_all = TRUE
)
}
\arguments{
\item{grid_size}{Int >0. Number of rows in the grid. Population size will be
the squre of this number.}

\item{mean_dispersal_distance}{Float >0. Mean seed dispersal distance.
Sampling points are arranged at intervals of 1m, so this should be relative
to that.}

\item{outcrossing_rate}{Float between 0 and 1. Probability that an individual
is outcrossed.}

\item{n_generations}{Int >34. Number of generations to run the simulations.}

\item{n_starting_genotypes}{Int >0. Number of initial genotypes to start with.
Defaults to 50}

\item{dormancy_rate}{Float >0. Strength of the seedbank. This is the rate
parameter for the poisson distribution. Giving zero will simulate no seed
bank.}

\item{verbose}{Logical. If TRUE, messages about simulation progress will be
printed. Defaults to TRUE.}

\item{return_all}{Logical. If TRUE, the entire population history will be
returned as a list. If FALSE, only the final generation is returned as a
vector. Defaults to TRUE.}
}
\value{
If `return_all` is TRUE, genotypes over the entire population history
will be returned as a list. If FALSE, only genotypes in the final generation
is returned as a vector.
}
\description{
Simulate a single population through time.
}
\details{
`sim_population` simulates a population on an evenly-spaced grid of sampling
points through time:

\enumerate{
   \item{Sampling points are populated at random with a set of unique
   starting genotypes.}
   \item{In the next generation, each sampling point is filled by one seed
   from the same or a different sampling point from a previous generation.
   Whichever genotype was at the donor sampling point now occupies the focal
   sampling point. I assume seed dispersal distances are exponentially
   distributed with some mean value that is used as an input parameter.}
   \item{To allow for a seed bank, each seed can be drawn from any previous
   generation. The generation is drawn from a poisson distribution.}
   \item{Each seed has some probability of having been the product of an
   outcrossing event. If so, it is assigned a new unique genotype.}
}

Because the total number of sampling points increases as the square of grid
width, the number of individuals is the limiting factor on simulations.
To speed things up `sim_population` first draws dispersal distances for
individual sampling points for all generations at once, and then fills in
the genotypes in a second step.
This is valid if we assume no differences in fitness or dispersal between
genotypes or though time (at least on average).
}
\author{
Tom Ellis
}
