% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_simmiad.R
\name{summarise_simmiad}
\alias{summarise_simmiad}
\title{Summarise output from \code{simmiad}}
\usage{
summarise_simmiad(base_folder, parameter)
}
\arguments{
\item{base_folder}{String giving the folder containing simulation results.
This folder should contain subfolders for each combination of parameters run,
each with a set of CSV files for that analysis.}

\item{parameter}{String giving the output parameter to be summarised (i.e.
the output CSV file to be opened).}
}
\value{
This returns a list giving means and CIs for the final generation, plus
means through time across generations.
}
\description{
Function to loop through folders and summarise simulation results over many
parameter combinations.
}
\author{
Tom Ellis
}
