% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_clustering.R
\name{transect_clustering}
\alias{transect_clustering}
\title{Estimate spatial clustering}
\usage{
transect_clustering(genotypes, positions)
}
\arguments{
\item{genotypes}{Vector of genotype labels}

\item{positions}{Vector of positions along a linear transect. Should be the
same length as \code{genotypes}.}
}
\value{
A vector of three elements giving the number of identical gentotypes
in the transect, the mean distance between sampling points with identical
genotypes and the mean distance between non-identical genotypes.
}
\description{
Estimate the degree of spatial clustering along a transect
}
\details{
\code{transect_clustering} estimates the degree of non-randomness in the
distribution of genotypes along a linear transect by comparing the mean
pairwise distances between sampling points with the same genotype to mean
pairwise distances between sampling points with non-identical genotypes.
}
\author{
Tom Ellis
}
