% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_stability.R
\name{transect_stability}
\alias{transect_stability}
\title{Changes in composition of two transects}
\usage{
transect_stability(x, y)
}
\arguments{
\item{x, y}{Vectors of genotypes along a transect.}
}
\value{
A number between 0 and 1. 0 indicates that no sampling point is
occupied by the same genotype, 1 indicates that every sampling point is
occupied by the same genotype.
}
\description{
Calculate the number of sampling points occupied by identical genotypes at
two time points.
}
\author{
Tom Ellis
}
