% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_table.R
\name{parameter_table}
\alias{parameter_table}
\title{Data.frame of parameter values}
\usage{
parameter_table(
  mean_dispersal_distance,
  outcrossing_rate,
  n_generations,
  n_starting_genotypes,
  density,
  nsims,
  n_sample_points,
  sample_spacing,
  range_limit,
  how_far_back,
  habitat_width = (n_sample_points * sample_spacing * range_limit),
  population_size = density * habitat_width^2
)
}
\arguments{
\item{mean_dispersal_distance}{Float >0. Mean seed dispersal distance. The
reciprocal of this is used as the rate parameter to draw from the exponential
distribution.}

\item{outcrossing_rate}{Float between 0 and 1. Probability that an individual
is outcrossed.}

\item{n_generations}{Int >12. Number of generations to run the simulations.}

\item{n_starting_genotypes}{Int >0. Number of initial genotypes to start with.
Defaults to 50}

\item{density}{Float >0. Average density of plants per square metre.}

\item{nsims}{Int >0. Number of replicate populations to simulate.}

\item{n_sample_points}{Number of points to sample along the transect.}

\item{sample_spacing}{Distance between sampling points.}

\item{range_limit}{Float >1 defining how much wider than the transect the
width of the habitat should be. This, along with plant density, determines
how many plants will be simulated. Defaults to 1.5.}

\item{how_far_back}{Integer number of generations to sample with a transect}

\item{habitat_width}{Length of the edge of the habitat square}

\item{population_size}{Number of individual plants in the population,
derived from the habitat size and population density}
}
\value{
A data.frame giving parameter names and values
}
\description{
Data.frame of parameter values
}
\author{
Tom Ellis
}
