% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_population.R
\name{sim_population}
\alias{sim_population}
\title{Simulate a population}
\usage{
sim_population(
  mean_dispersal_distance,
  outcrossing_rate,
  n_generations,
  n_starting_genotypes,
  density,
  n_sample_points,
  sample_spacing,
  range_limit = 1.5
)
}
\arguments{
\item{mean_dispersal_distance}{Float >0. Mean seed dispersal distance. The
reciprocal of this is used as the rate parameter to draw from the exponential
distribution.}

\item{outcrossing_rate}{Float between 0 and 1. Probability that an individual
is outcrossed.}

\item{n_generations}{Int >12. Number of generations to run the simulations.}

\item{n_starting_genotypes}{Int >0. Number of initial genotypes to start with.
Defaults to 50}

\item{density}{Float >0. Average density of plants per square metre.}

\item{n_sample_points}{Number of points to sample along the transect.}

\item{sample_spacing}{Distance between sampling points.}

\item{range_limit}{Float >1 defining how much wider than the transect the
width of the habitat should be. This, along with plant density, determines
how many plants will be simulated. Defaults to 1.5.}
}
\value{
A list of genotypes recorded at each sampling point in each
generation.
}
\description{
Simulate a single population through time on a torus, and sample a transect
for each generation.
}
\details{
\code{sim_population} simulates a population of annuals dispersing in continuous
space following exponentially distributed seed dispersal distances.
\enumerate{
\item Individuals are drawn from a set of starting genotypes and randomly
distributed in a square habitat Individuals are labelled by their genotype
as 'g' followed by an integer label.
\item Individuals are chosen at random to found the next generation. They
move seed dispersal to new location, with distances drawn from an exponential
distribution. The population exists on a torus to eliminat edge effects;
if dispersal takes a plant of the edge of the range it moves to the other
side of the range.
\item A subset of the new generation are chosen to have been germinated from
outcrossed seed. If so, they are given a new unique label by appending their
current genotype label by the generation number plus a unique integer.
\item At each generation a transect through the population is drawn at x=0 with
sampling points at equally spaced intervals. The plant closest to each
sampling point is recorded.
}

A seed bank is not currently implemented.

Plants exist within a box centred around zero with a transect going through
zero. The width of the box is defined to be 1.5-fold larger than the length
of the transect (but this can be changed with the argument \code{range_limit}).
Total population size is then the density of plants multiplied by the squared
width of the box.
}
\author{
Tom Ellis
}
