% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_population.R
\name{sim_population}
\alias{sim_population}
\title{Simulate a population}
\usage{
sim_population(
  grid_size,
  mean_dispersal_distance,
  outcrossing_rate,
  n_generations,
  n_starting_genotypes = 124,
  verbose = TRUE,
  return_all = FALSE
)
}
\arguments{
\item{grid_size}{Int >0. Number of rows in the grid. Population size will be
the squre of this number.}

\item{mean_dispersal_distance}{Float >0. Mean seed dispersal distance.
Sampling points are arranged at intervals of 1m, so this should be relative
to that.}

\item{outcrossing_rate}{Float between 0 and 1. Probability that an individual
is outcrossed.}

\item{n_generations}{Int >34. Number of generations to run the simulations.}

\item{n_starting_genotypes}{Int >0. Number of initial genotypes to start with.
Defaults to 124.}

\item{verbose}{Logical. If TRUE, messages about simulation progress will be
printed. Defaults tp}

\item{return_all}{Logical. If TRUE, the entire population history will be
returned as a list. If FALSE, only the final generation is returned as a
vector.}
}
\value{
If `return_all` is TRUE, genotypes over the entire population history
will be returned as a list. If FALSE, only genotypes in the final generation
is returned as a vector.
}
\description{
Simulate a single population through time.
}
\details{
This sets up a population with \eqn{N} squared sampling points arranged at 1m
intervals on a grid.
Sampling points are then populated by randomly assigning one of a set of
starting genotypes to each starting point (i.e. there is no initial spatial
structure).
In the next generation, the genotype at each sampling point is filled by seed
dispersal, assuming dispersal distances are exponentially distributed.
A single donor plant is chosen from the previous generation in proportion to
the probability of seeds travelling that distance under the exponential.
To account for outcrossing, each establishing seed is randomly assigned to be
outcrossed seed or not in proportion to a given outcrossing rate.
Outcrossed seeds are given a new unique genotype.

Because the total number of sampling points increases as the square of grid
width, the number of individuals is the limiting factor on simulations.
To speed things up `sim_population` first draws dispersal distances for
individual sampling points for all generations at once, and then fills in
the genotypes in a second step.
This is valid if we assume no differences in fitness or dispersal between
genotypes or though time (at least on average).
}
\author{
Tom Ellis
}
