% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmiad.R
\name{simmiad}
\alias{simmiad}
\title{Simulate replicate populations}
\usage{
simmiad(
  grid_size,
  mean_dispersal_distance,
  outcrossing_rate,
  n_generations,
  n_starting_genotypes = 124,
  filename = "simmiad",
  nsims
)
}
\arguments{
\item{grid_size}{Int >0. Number of rows in the grid. Population size will be
the squre of this number.}

\item{mean_dispersal_distance}{Float >0. Mean seed dispersal distance.
Sampling points are arranged at intervals of 1m, so this should be relative
to that.}

\item{outcrossing_rate}{Float between 0 and 1. Probability that an individual
is outcrossed.}

\item{n_generations}{Int >0. Number of generations to run the simulations.}

\item{n_starting_genotypes}{Int >0. Number of initial genotypes to start with.
Defaults to 124.}

\item{filename}{Str. Directory where output and log file should be saved.}

\item{nsims}{Int >0. Number of replicate populations to simulate.}
}
\value{
Nothing will be printed on screeen, but two files are saved to disk:
A CSV file giving genotype distances between identical and
non-identical plants, and a log file giving simulation details.
}
\description{
`simmiad` Simulates multiple replicate populations under a set of input
parameters, and outputs a summary of the clustering among genotypes.
}
\details{
`simmiad` is a wrapper function to call multiple instances of `sim_population`,
estimate the degree of clustering in each, and save results to disk. See the
help file `?sim_population` for details of individual population simulations.

In the final generation of each population, a single transect is taken through
a random row of the grid, and the degree of clustering estimated using
`transect_cluster`. This summarises the mean distance between pairs of
identical genotypes and between pairs of non-identical genotypes. Summary data
are saved to disk in a CSV file.

Need to add sth about parallelising.
}
\seealso{
`sim_population`, `transect_cluster`
}
\author{
Tom Ellis
}
