% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmiad.R
\name{simmiad}
\alias{simmiad}
\title{Simulate replicate populations}
\usage{
simmiad(
  population_size,
  mean_dispersal_distance,
  outcrossing_rate,
  n_generations,
  n_starting_genotypes = 50,
  density = 3,
  n_sample_points = 30,
  sample_spacing = 5,
  nsims,
  progress = TRUE,
  how_far_back = n_generations
)
}
\arguments{
\item{mean_dispersal_distance}{Float >0. Mean seed dispersal distance. The
reciprocal of this is used as the rate parameter to draw from the exponential
distribution.}

\item{outcrossing_rate}{Float between 0 and 1. Probability that an individual
is outcrossed.}

\item{n_generations}{Int >12. Number of generations to run the simulations.}

\item{n_starting_genotypes}{Int >0. Number of initial genotypes to start with.
Defaults to 50}

\item{density}{Float >0. Average density of plants per square metre.}

\item{n_sample_points}{Number of points to sample along the transect.}

\item{sample_spacing}{Distance between sampling points.}

\item{nsims}{Int >0. Number of replicate populations to simulate.}

\item{progress}{If TRUE, a progress bar is printed.}

\item{how_far_back}{Integer number of generations}
}
\value{
A data.frame giving simulation replicate, number of identical and
non-identical genotypes in the transect, mean distances between identical
and non-identical plants and the output of transect_stability for each pair
of sampling years.
}
\description{
\code{simmiad} Simulates multiple replicate populations under a set of input
parameters, and outputs a summary of the clustering among genotypes.
}
\details{
\code{simmiad} is a wrapper function to call multiple instances of \code{sim_population},
estimate the degree of clustering in each, and save results to disk. See the
help file \code{?sim_population} for details of individual population simulations.

For each simulation, \code{simmiad} draws a horizontal transect and saves
two kinds of summary information about this transect to disk. First,
in the final generation of each population, a single transect is taken through
a random row of the grid, and the degree of clustering estimated using
\code{transect_cluster}. This summarises the number and mean distance between pairs of
identical genotypes and between pairs of non-identical genotypes.
Second, it estimates the spatial stability of the populations by calculating
how often a sampling point is occupied by idnetical genotypes at different
time points. This is done by comparing the final generation to 1, 2, 4, 6 and
twelve years prior to that.
}
\seealso{
\code{sim_population}, \code{transect_cluster}, \code{transect_stability}
}
\author{
Tom Ellis
}
