% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmiad.R
\name{simmiad}
\alias{simmiad}
\title{Simulate replicate populations}
\usage{
simmiad(
  mean_dispersal_distance,
  outcrossing_rate,
  n_generations,
  n_starting_genotypes = 50,
  density = 3,
  n_sample_points = 30,
  sample_spacing = 5,
  range_limit = 1.5,
  nsims,
  progress = TRUE,
  dormancy,
  how_far_back = n_generations,
  stability_years = n_generations:(n_generations/2),
  pop_structure = "uniform",
  mixing = mean_dispersal_distance
)
}
\arguments{
\item{mean_dispersal_distance}{Float >0. Mean seed dispersal distance. The
reciprocal of this is used as the rate parameter to draw from the exponential
distribution.}

\item{outcrossing_rate}{Float between 0 and 1. Probability that an individual
is outcrossed.}

\item{n_generations}{Int >12. Number of generations to run the simulations.}

\item{n_starting_genotypes}{Int >0. Number of initial genotypes to start with.
Defaults to 50}

\item{density}{Float >0. Average density of plants per square metre.}

\item{n_sample_points}{Number of points to sample along the transect.}

\item{sample_spacing}{Distance between sampling points.}

\item{range_limit}{Float >1 defining how much wider than the transect the
width of the habitat should be. This, along with plant density, determines
how many plants will be simulated. Defaults to 1.5. Ignored if
\code{pop_structure="hardcoded"}

#' @param pop_structure Character string indicating the initial population
structure to be simulated. Passing "uniform" simulated a panmictic population.
"clusters" simulates clusters of identical individuals that disperse from
distinct mothers via exponential dispersal set by \code{mean_dispersal_distance}.
This is likely to generate very disperate clumps. Passing "mvnorm" simulates
uniformly distributed coordinates for indiduals, as well as centroid
positions for genotypes. Individuals are assigned a genotype in proportion to
their distance to each genotype centroid based on multivariate-normal
probabilities. The variance covariance matrix for this is set as
\code{sqrt(habitat_size/n_starting_genotypes) / 3} such that the tail of each
genotype just about touches those of its neighbours, on average.
If \code{pop_structure='hardcoded'} and a vector of genotypes is passed to
\code{n_starting_genotypes}, for example
observed genotypes from along all real-world transects, this simulates
bands of identical genotypes by copying the vector over an evenly-
spaced grid (giving horizontal but not vertical structure). There is one
round of dispersal from this initial generation via exponential dispersal
(controlled by \code{mixing}) and to get the population to the correct
population density.}

\item{nsims}{Int >0. Number of replicate populations to simulate.}

\item{progress}{If TRUE, a progress bar is printed.}

\item{dormancy}{Float between 0 and 1. Probability that a seedling is drawn
from the seed bank. Seedlings are drawn from the prior generation with
probability 1-dormancy.}

\item{how_far_back}{Integer number of generations to sample with a transect}

\item{stability_years}{Vector of integers indexing which generations to average
over when calculating distance_identities(). Defaults to the last 50\% of
generations.}

\item{mixing}{Float >0. Parameter controlling the degree of spatial
clustering of genotypes. Smaller values indicate more structure populations.
If \code{pop_structure='mvnorm'} this is a scaler multiplier for the variance
of the multivariate normal probability density.
If \code{pop_structure="clusters`} or \code{pop_structure='hardcoded'} this is
the reciprocal of the rate parameter to draw dispersal distances from the
exponential distribution.}
}
\value{
A list of seven dataframes.
\enumerate{
\item \strong{parameters} A data.frame giving input parameters.
\item \strong{clustering} The covariance between distance along the transect and the
frequency of identical genotypes.
\item \strong{matching_pairs}: The number of pairs of identical genotypes in the
transect.
\item \strong{count_NA}: The number of empty sampling points.
\item \strong{n_genotypes}: The number of unique genotypes sampled in the transect
(note that this will be different from what you gave as \code{distance_identity},
because the latter reflects genotypes in \emph{the whole population}, not just in
the transect).
\item \strong{stability}: How often individual sampling points are occupied by the
same genotype in the final generations and 1, 2, ..., n generations back.
\item \strong{distance_identity}: Probabilities of finding identical genotypes in
pairs of sampling points at all possible distances between transects. For
example, if there are five evenly spaced sampling points as in the example
above, there are four possible distances between sampling points. Rows
indicate replicate simulations.
}

For 2-6, rows indicate replicate simulations and columns generations.
}
\description{
\code{simmiad} Simulates multiple replicate populations under a set of input
parameters, and outputs a summary of the clustering among genotypes.
}
\details{
\code{simmiad} is a wrapper function to call multiple instances of \code{sim_population},
estimate the degree of clustering in each, and save results to disk. See the
help file \code{?sim_population} for details of individual population simulations.

For each simulation, \code{simmiad} draws a horizontal transect and saves
two kinds of summary information about this transect to disk. First,
in the final generation of each population, a single transect is taken through
a random row of the grid, and the degree of clustering estimated using
\code{transect_cluster}. This summarises the number and mean distance between pairs of
identical genotypes and between pairs of non-identical genotypes.
Second, it estimates the spatial stability of the populations by calculating
how often a sampling point is occupied by idnetical genotypes at different
time points. This is done by comparing the final generation to 1, 2, 4, 6 and
twelve years prior to that.
}
\seealso{
\code{sim_population}, \code{transect_cluster}, \code{transect_stability}
}
\author{
Tom Ellis
}
