% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_simmiad.R
\name{write_simmiad}
\alias{write_simmiad}
\title{Write to disk}
\usage{
write_simmiad(x, directory)
}
\arguments{
\item{x}{Directory where output should be saved.}
}
\description{
Write the output of simmiad simulations to disk
}
\author{
Tom Ellis
}
